REF - Refactor all route, store and provider configuration into top level App component. Removed separate UI reduers, action creators as App + Router can be used to store top level UI state.
